fix indexer for kudu and waitForResults was definitely bugged on the client
