+++ src/client/utils/search-utils.ts
@@ -8,8 +8,21 @@ import { runScriptsInContainer } from "./search-helpers";
 
 let glanceAbortController: AbortController | null = null;
 let slotsAbortController: AbortController | null = null;
+let independentKnowledgePanels: SlotPanel[] = [];
 
 const _skipSlotPanels = (type: string): boolean => isImageSearchType(type);
+
+const _knowledgePanels = (panels: SlotPanel[]): SlotPanel[] =>
+  panels.filter((p) => p.position === SlotPanelPosition.KnowledgePanel);
+
+const _mergeKnowledgePanels = (panels: SlotPanel[]): SlotPanel[] => {
+  const dependent = _knowledgePanels(panels);
+  const seen = new Set(dependent.map((p) => p.id));
+  return [
+    ...independentKnowledgePanels.filter((p) => !seen.has(p.id)),
+    ...dependent,
+  ];
+};
 
 export const abortGlancePanels = (): void => {
   if (glanceAbortController) {
@@ -23,52 +36,74 @@ export const abortSlotPanels = (): void => {
     slotsAbortController.abort();
     slotsAbortController = null;
   }
+  independentKnowledgePanels = [];
+};
+
+const _slotRequestBody = (
+  query: string,
+  results?: ScoredResult[],
+): string =>
+  JSON.stringify(
+    results !== undefined
+      ? { query: query.trim(), results }
+      : { query: query.trim() },
+  );
+
+const _renderGlanceHtml = (panels: SlotPanel[], clearIfEmpty: boolean): void => {
+  const glanceEl = document.getElementById("at-a-glance");
+  if (!glanceEl) return;
+  const glancePanels = panels.filter(
+    (p) => p.position === SlotPanelPosition.AtAGlance,
+  );
+  if (glancePanels.length === 0) {
+    if (clearIfEmpty) glanceEl.innerHTML = "";
+    return;
+  }
+  const parts: string[] = [];
+  for (const panel of glancePanels) {
+    const titleHtml = panel.title
+      ? `<div class="results-slot-panel-title degoog-panel--slot-title">${escapeHtml(panel.title)}</div>`
+      : "";
+    parts.push(
+      `<div class="results-slot-panel degoog-panel degoog-panel--slot degoog-panel--stack-item">${titleHtml}<div class="results-slot-panel-body degoog-panel--slot-body degoog-panel--slot-body-padded">${panel.html}</div></div>`,
+    );
+  }
+  glanceEl.innerHTML = parts.join("");
+  runScriptsInContainer(glanceEl);
 };
 
 export async function fetchGlancePanels(
   query: string,
-  results: ScoredResult[],
+  results?: ScoredResult[],
 ): Promise<void> {
-  abortGlancePanels();
-  glanceAbortController = new AbortController();
-  const signal = glanceAbortController.signal;
-  const glanceEl = document.getElementById("at-a-glance");
-  if (!results || results.length === 0) {
+  if (results !== undefined && results.length === 0) {
+    abortGlancePanels();
+    const glanceEl = document.getElementById("at-a-glance");
     if (glanceEl) glanceEl.innerHTML = "";
     return;
   }
+  if (results === undefined) {
+    abortGlancePanels();
+    glanceAbortController = new AbortController();
+  } else if (!glanceAbortController) {
+    glanceAbortController = new AbortController();
+  }
+  const signal = glanceAbortController!.signal;
   try {
     const res = await fetch(`${getBase()}/api/slots/glance`, {
       method: "POST",
       headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ query: query.trim(), results }),
+      body: _slotRequestBody(query, results),
       signal,
     });
     if (signal.aborted) return;
     const data = (await res.json()) as { panels?: SlotPanel[] };
     if (signal.aborted) return;
     if (_skipSlotPanels(state.currentType)) return;
-    if (!glanceEl) return;
-    if (data.panels && data.panels.length > 0) {
-      const glancePanels = data.panels.filter(
-        (p) => p.position === SlotPanelPosition.AtAGlance,
-      );
-      const parts: string[] = [];
-      for (const panel of glancePanels) {
-        const titleHtml = panel.title
-          ? `<div class="results-slot-panel-title degoog-panel--slot-title">${escapeHtml(panel.title)}</div>`
-          : "";
-        parts.push(
-          `<div class="results-slot-panel degoog-panel degoog-panel--slot degoog-panel--stack-item">${titleHtml}<div class="results-slot-panel-body degoog-panel--slot-body degoog-panel--slot-body-padded">${panel.html}</div></div>`,
-        );
-      }
-      glanceEl.innerHTML = parts.join("");
-      runScriptsInContainer(glanceEl);
-    } else {
-      if (glanceEl) glanceEl.innerHTML = "";
-    }
+    _renderGlanceHtml(data.panels ?? [], results !== undefined);
   } catch (err) {
     if (err instanceof Error && err.name === "AbortError") return;
+    const glanceEl = document.getElementById("at-a-glance");
     if (glanceEl) glanceEl.innerHTML = "";
   }
 }
@@ -77,14 +112,18 @@ export async function fetchSlotPanels(
   query: string,
   results?: ScoredResult[],
 ): Promise<SlotPanel[]> {
-  abortSlotPanels();
-  slotsAbortController = new AbortController();
-  const signal = slotsAbortController.signal;
+  if (results === undefined) {
+    abortSlotPanels();
+    slotsAbortController = new AbortController();
+  } else if (!slotsAbortController) {
+    slotsAbortController = new AbortController();
+  }
+  const signal = slotsAbortController!.signal;
   try {
     const res = await fetch(`${getBase()}/api/slots`, {
       method: "POST",
       headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({ query: query.trim(), results }),
+      body: _slotRequestBody(query, results),
       signal,
     });
     if (signal.aborted) return [];
@@ -93,10 +132,16 @@ export async function fetchSlotPanels(
     if (signal.aborted) return [];
     if (_skipSlotPanels(state.currentType)) return [];
     const panels = data.panels ?? [];
+    if (results === undefined) {
+      independentKnowledgePanels = _knowledgePanels(panels);
+      if (panels.length > 0) appendSlotPanels(panels);
+      return panels;
+    }
+    const merged = _mergeKnowledgePanels(panels);
     if (panels.length > 0) appendSlotPanels(panels);
-    return panels;
+    return merged;
   } catch {
-    return [];
+    return results === undefined ? [] : _mergeKnowledgePanels([]);
   }
 }
