+++ src/server/indexer/adapters/index.ts
@@ -1,0 +1,2 @@
+export { SqliteAdapter } from "./sqlite/adapter";
+export { PgAdapter } from "./postgres/adapter";
