+++ src/server/indexer/db/index.ts
@@ -1,0 +1,2 @@
+export { getAdapter, bootAdapter, isPostgresMode } from "./factory";
+export { discoverTypes, checkpointType, closeAllDbs } from "./lifecycle";
