+++ src/server/indexer/filters/index.ts
@@ -1,0 +1,1 @@
+export { shouldIndex } from "./filters";
