+++ src/server/indexer/store/index.ts
@@ -1,0 +1,9 @@
+export { checkpointType } from "../db/lifecycle";
+export type { ExportRow, HitRow } from "../types/adapter";
+export type { DeleteItem, ServerIndexerStats as IndexerStats } from "../../../shared/indexer";
+export { DEGOOG_ENGINE_NAME } from "./mapper";
+export { recordResults, maybeIndex } from "./record";
+export { queryIndex } from "./query";
+export { getStats, wipeStatsCache } from "./stats";
+export { listHits, countHits, deleteHits } from "./hits";
+export { getKnownTypes, clearAll, sampleRows } from "./admin";
