+++ src/server/types/indexer.ts
@@ -1,0 +1,1 @@
+export type { DeleteItem, ServerIndexerStats } from "../../shared/indexer";
